Expand all queries (multiple pages)
Get a flat list of subqueries.
For example, from a query with "page": "+3", 3 queries with "page": "1",
"page": "2" and "page": "3" will be expanded.
This is so we can handle all requests seamlessly with multiprocessing
later, different queries and subqueries for each pages.

Also use generators wherever possible, for lazy handling and
performance.