Accurate artcom checking, no more has_notes/artcom
A request will be made if:
the post has a "commentary" or "commentary_request" tag,
or if it was posted in the 24 hours.

With this design has_artcom() becomes innacurate, if we put the request check
in it in the end we'll do yet another request to really retrive the artcom.

Now there is just one function to get the artcom, if there's no artcom
(the request wasn't done, or the request returned []), the function returns [].

To have the same consistent design, the has_notes() was removed too.

download.py will not write a file if the function return is [].