Error and CTRL-C protection for Album.put
Catch any unexpected exception that could happen for every post,
show traceback, and continue trying to finish the task;
to prevent random broken post or unknown bugs from destroying
a long-running user search.

On CTRL-C, cancel, all posts that had time to be put in the Album will
be there and not discarded.