Config system and auto-filter
Auto-created clients will now be based on the user config (a default one
including Danbooru and Safebooru is copied if it doesn't exist).

The config currently allows to define:
Site name/url to create clients for, login details,
max number of parallel requests, and an automatic tag filter that can be
used like a blacklist.